Projects/rag_app/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
pydantic

langchain
langchain-community
langchain-text-splitters

chromadb
pypdf

sentence-transformers
torch

# Optional LLMs
google-generativeai
ollama
